+++ packages/inline-choice/configure/__mocks__/@pie-libs/config-ui.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-module.exports = {
-  MultiLangInput: () => <div>mui</div>
-};