+++ packages/inline-choice/controller/src/__test__/index.test.js
@@ -1,8 +1,6 @@
 import { model } from '../index';
 
 describe('model', () => {
-
-
   let result;
 
   let question = {
@@ -25,20 +23,22 @@ describe('model', () => {
         }
       }
     ]
-  }
+  };
 
   let session = {
-    selectedChoice: 'a'
-  }
+    value: 'a'
+  };
 
   let env;
   beforeEach(async () => {
-    env = { mode: 'gather' }
+    env = { mode: 'gather' };
     result = await model(question, session, env);
   });
 
   it('returns choices', () => {
-    expect(result.choices).toMatchObject(question.choices.map(c => ({ label: c.label, value: c.value })));
+    expect(result.choices).toMatchObject(
+      question.choices.map(c => ({ label: c.label, value: c.value }))
+    );
   });
 
   it('returns disabled:false', () => {
@@ -46,7 +46,6 @@ describe('model', () => {
   });
 
   describe('mode == evaluate', () => {
-
     beforeEach(async () => {
       env = { mode: 'evaluate' };
       result = await model(question, session, env);
@@ -60,17 +59,34 @@ describe('model', () => {
       expect(result.result).toMatchObject({
         correct: true,
         feedback: 'hooray'
-      })
+      });
+    });
+  });
+
+  describe('mode === evaluate nothing submitted', () => {
+    beforeEach(async () => {
+      env = { mode: 'evaluate' };
+      session = {
+        value: undefined
+      };
+      result = await model(question, session, env);
+    });
+
+    it('returns result', () => {
+      expect(result.result).toMatchObject({
+        correct: false,
+        nothingSubmitted: true,
+        feedback: undefined
+      });
     });
   });
 
   describe('mode === evaluate wrong answer', () => {
-
     beforeEach(async () => {
-      env = { mode: 'evaluate' }
+      env = { mode: 'evaluate' };
       session = {
-        selectedChoice: 'b'
-      }
+        value: 'b'
+      };
       result = await model(question, session, env);
     });
 
@@ -79,6 +95,6 @@ describe('model', () => {
         correct: false,
         feedback: 'Incorrect'
       });
-    })
+    });
   });
-})
+});