+++ packages/inline-choice/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.7.1"></a>
+## [1.7.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@1.7.0...@pie-element/inline-choice@1.7.1) (2018-04-25)
+
+
+### Bug Fixes
+
+* **dependency:** bump [@pie-ui](https://github.com/pie-ui)/inline-choice to ^4.0.1 ([af6c79d](https://github.com/pie-framework/pie-elements/commit/af6c79d))
+
+
+
+
+<a name="1.7.0"></a>
+# [1.7.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-choice@1.6.0...@pie-element/inline-choice@1.7.0) (2018-04-25)
+
+
+### Features
+
+* **layout:** use [@pie-ui](https://github.com/pie-ui)/inline-choice@^4.0.1 - it uses an inline layout ([b91827f](https://github.com/pie-framework/pie-elements/commit/b91827f)), closes [#17](https://github.com/pie-framework/pie-elements/issues/17)
+
+
+
+
+<a name="1.6.0"></a>
+# 1.6.0 (2018-04-25)
+
+
+### Features
+
+* **root:** use [@pie-lib](https://github.com/pie-lib)/config-ui ([974ec8d](https://github.com/pie-framework/pie-elements/commit/974ec8d))
+* **value:** remove 'value' input, it is now generated based on the choice index. ([a1459fc](https://github.com/pie-framework/pie-elements/commit/a1459fc))
+
+
+
+<a name="1.5.0"></a>
+# 1.5.0 (2018-04-17)
+
+
+### Bug Fixes
+
+* **dependencies:** update dependencies ([c774453](https://github.com/pie-framework/pie-elements/commit/c774453))
+* **icon:** use [@material-ui](https://github.com/material-ui)/icons package ([9e45c1a](https://github.com/pie-framework/pie-elements/commit/9e45c1a))
+
+
+
+<a name="1.2.2"></a>
+## 1.2.2 (2018-03-06)
+
+
+### Bug Fixes
+
+* **dependencies:** fix dependencies ([5ea0806](https://github.com/pie-framework/pie-elements/commit/5ea0806))
+* **dependencies:** update dependency names ([dc789ac](https://github.com/pie-framework/pie-elements/commit/dc789ac))
+
+
+
+
 <a name="1.5.0"></a>
 # [1.5.0](https://github.com/pie-framework/pie-elements/compare/v1.4.1...v1.5.0) (2018-04-17)